fix: add average

+++ cmc-totaler/src/main.rs
@@ -50,8 +50,8 @@ fn main() {
         }
     }
 
-    let cards_in_deck = 100; // default to cmder deck for now
-    let cmc_average = total / cards_in_deck;
+    let cards_in_deck = 100.0; // default to cmder deck for now
+    let cmc_average = total as f32 / cards_in_deck;
 
     println!("\nTotal CMC count: {}", total);
     println!("CMC Average: {}", cmc_average);
